refactor : iniComponent 상수로 관리
