+++ src/domain/index.js
@@ -1,3 +1,4 @@
 export { default as Document } from "./Document.js";
 export { default as DocumentTree } from "./DocumentTree.js";
 export { default as DocumentTreeBranch } from "./DocumentTreeBranch.js";
+export { initDocument, initComponent } from "./initData.js";
