+++ src/events/documentTreeEvent.js
@@ -3,9 +3,8 @@ import {
   removeDocumentFromStorage,
 } from "../service/index.js";
 import { hashRouter } from "../router/hashRouter.js";
-import { Document } from "../domain/index.js";
+import { Document, initDocument } from "../domain/index.js";
 import { request } from "../api.js";
-import { initDocument } from "../constant/constants.js";
 
 export const documentLinkClickEvent = async ({ url, event }) => {
   event.preventDefault();
